+++ lib/src/extension_color.dart
@@ -2,21 +2,10 @@ part of '../ultimate_extension.dart';
 
 extension ColorHelper on Color {
   ///To create a Color object from a hexadecimal string:
-  Color fromHex(String hexString) {
-    final buffer = StringBuffer();
-    if (hexString.length == 6 || hexString.length == 7) buffer.write('ff');
-    buffer.write(hexString.replaceFirst('#', ''));
-    return Color(int.parse(buffer.toString(), radix: 16));
-  }
+  
 
   /// To generate a random Color:
-  Color getRandomColor() {
-    final Random random = Random();
-    final int r = random.nextInt(256);
-    final int g = random.nextInt(256);
-    final int b = random.nextInt(256);
-    return Color.fromARGB(255, r, g, b);
-  }
+
 
   /// Generates a random shade of the current color.
   Color getRandomShade() {